requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Computer Vision & AI
opencv-python==4.8.1.78
torch==2.7.1
torchvision==0.19.1
ultralytics==8.0.200
mediapipe==0.10.7
timm==0.9.10
transformers==4.35.2

# Video Processing
moviepy==1.0.3
imageio==2.33.0
imageio-ffmpeg==0.4.9
Pillow==10.1.0

# Tracking & Analysis
lap==0.4.0
filterpy==1.4.5
scipy==1.11.4
scikit-learn==1.3.2
networkx==3.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.18.0

# API & Async
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4

# Utils
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Documentation & Reports
reportlab==4.0.7
jinja2==3.1.2
markdown==3.5.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0